move NumpyFeature to its own module
