+++ temporian/implementation/numpy/data/sampling.py
@@ -15,9 +15,21 @@ class NumpySampling:
         data: Dict[Tuple, np.ndarray],
         is_unix_timestamp: bool = False,
     ) -> None:
-        self.index = index
-        self.data = data
-        self.is_unix_timestamp = is_unix_timestamp
+        self._index = index
+        self._data = data
+        self._is_unix_timestamp = is_unix_timestamp
+
+    @property
+    def index(self) -> List[str]:
+        return self._index
+
+    @property
+    def data(self) -> Dict[Tuple, np.ndarray]:
+        return self._data
+
+    @property
+    def is_unix_timestamp(self) -> bool:
+        return self._is_unix_timestamp
 
     @property
     def has_repeated_timestamps(self) -> bool:
