+++ temporian/implementation/numpy/operators/calendar/BUILD
@@ -17,9 +17,10 @@ py_library(
     srcs = ["base.py"],
     srcs_version = "PY3",
     deps = [
-        "//temporian/implementation/numpy/operators:base",
         "//temporian/core/operators/calendar:base",
         "//temporian/implementation/numpy/data:event",
+        "//temporian/implementation/numpy/data:feature",
+        "//temporian/implementation/numpy/operators:base",
     ],
 )
