+++ temporian/implementation/numpy/operators/test/calendar_hour_test.py
@@ -20,7 +20,7 @@ from temporian.core.operators.calendar.hour import (
     CalendarHourOperator,
 )
 from temporian.implementation.numpy.data.event import NumpyEvent
-from temporian.implementation.numpy.data.event import NumpyFeature
+from temporian.implementation.numpy.data.feature import NumpyFeature
 from temporian.implementation.numpy.operators.calendar.hour import (
     CalendarHourNumpyImplementation,
 )
@@ -65,7 +65,7 @@ class CalendarHourNumpyImplementationTest(absltest.TestCase):
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
-            output["event"]._first_index_features[0].dtype == dtype.INT32
+            output["event"].first_index_features()[0].dtype == dtype.INT32
         )
