+++ temporian/implementation/numpy/operators/test/calendar_minute_test.py
@@ -20,7 +20,7 @@ from temporian.core.operators.calendar.minute import (
     CalendarMinuteOperator,
 )
 from temporian.implementation.numpy.data.event import NumpyEvent
-from temporian.implementation.numpy.data.event import NumpyFeature
+from temporian.implementation.numpy.data.feature import NumpyFeature
 from temporian.implementation.numpy.operators.calendar.minute import (
     CalendarMinuteNumpyImplementation,
 )
@@ -66,7 +66,7 @@ class CalendarMinuteNumpyImplementationTest(absltest.TestCase):
 
         self.assertTrue(output_event_data == output["event"])
         self.assertTrue(
-            output["event"]._first_index_features[0].dtype == dtype.INT32
+            output["event"].first_index_features()[0].dtype == dtype.INT32
         )
