+++ temporian/implementation/numpy/operators/window/BUILD
@@ -17,10 +17,11 @@ py_library(
     srcs = ["base.py"],
     srcs_version = "PY3",
     deps = [
-        "//temporian/implementation/numpy/operators:base",
         "//temporian/core/data:duration",
         "//temporian/core/operators/window:base",
         "//temporian/implementation/numpy/data:event",
+        "//temporian/implementation/numpy/data:feature",
+        "//temporian/implementation/numpy/operators:base",
     ],
 )
